test(apollo-forest-run): move incomplete recycling test into regression.test.ts
